Combined logic for getting positions, renaming and moved migrations into separate file
