+++ core/src/util/mod.rs
@@ -4,5 +4,6 @@ pub mod db_helpers;
 pub mod general_helpers;
 pub mod import_helpers;
 pub mod market_data_helpers;
+pub mod migrations;
 pub mod performance_helpers;
 pub mod taxation_helpers;
