+++ web/src/pages/dashboard/Dashboard.tsx
@@ -1,11 +1,11 @@
 import { Tab, TabList, TabGroup, TabPanels, TabPanel } from "@tremor/react";
 import useAuth from "@/hooks/useAuth";
-import Positions from "@/pages/dashboard/Positions";
 import PL from "@/pages/dashboard/PL";
 import Dividends from "@/pages/dashboard/Dividends";
-import ActiveUnits from "@/pages/dashboard/ActiveUnits";
 import Taxation from "@/pages/dashboard/Taxation";
 import Timeline from "@/pages/dashboard/Timeline";
+import Positions from "@/pages/dashboard/Positions";
+import Portfolio from "@/pages/dashboard/Portfolio";
 
 const Dashboard = () => {
   const { logout } = useAuth();
@@ -27,9 +27,7 @@ const Dashboard = () => {
             <Tab className="flex-shrink-0" value={3}>
               Dividends
             </Tab>
-            <Tab className="flex-shrink-0" value={4}>
-              Active Units
-            </Tab>
+            <Tab className="flex-shrink-0" value={4}>Positions</Tab>
             <Tab className="flex-shrink-0" value={5}>
               Taxation
             </Tab>
@@ -40,7 +38,7 @@ const Dashboard = () => {
         </TabList>
         <TabPanels>
           <TabPanel>
-            <Positions />
+            <Portfolio />
           </TabPanel>
           <TabPanel>
             <PL />
@@ -52,7 +50,7 @@ const Dashboard = () => {
             <Dividends />
           </TabPanel>
           <TabPanel>
-            <ActiveUnits />
+            <Positions />
           </TabPanel>
           <TabPanel>
             <Taxation />
